Use ILIKE to prevent case-sensitivity issues
I notice that when using PostgreSQL, the user's login credential (email address, by default) is treated in a case-sensitive manner. So, joeblow@localhost.dev works, but JoeBlow@localhost.dev does not.

I'm not sure if the solution is as simple as what I'm proposing here, but I can't think of a scenario in which the username should ever be treated in a case-sensitive fashion (at least when email address is used as the primary user identifier; a username might be a different animal).